rewrite in AppleScriptObjC
To fix bug with the app opening when running and add support for supplying a bundle id.

Before this the app we used in the AppleScript was opened because AppleScript needed to read its dictionary which it apparently can only do when the app is opening. It now uses the Cocoa API instead which doesn't have this weird behaviour.