rewrite in Objective-C
At least I get to practise my Swift and Objective-C skills...

Turns out if you compile Swift using `swiftc` it won't embed the required Swift runtime. Without the runtime it won't work on Mavericks or maybe even not on future OS X versions. With the runtime, when compiled through Xcode, the binary is 2.5 MB which is way too large for my use-case...

http://blog.diogot.com/blog/2014/09/29/swift-embedded-runtime-library-increases-app-size/